git-svn-id: https://svn.sosy-lab.org/software/cpachecker/branches/bamRecursion@14456 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c

+++ src/org/sosy_lab/cpachecker/cpa/arg/ARGMergeJoinPredicatedAnalysis.java
@@ -38,11 +38,13 @@ import org.sosy_lab.cpachecker.exceptions.PredicatedAnalysisPropertyViolationExc
 
 public class ARGMergeJoinPredicatedAnalysis implements MergeOperator {
 
+  private final boolean deleteSubgraphAfterMerge;
   private final MergeOperator wrappedMerge;
-  private ArrayList<ARGState> toDeleteFromReached = new ArrayList<>();
+  private final ArrayList<ARGState> toDeleteFromReached = new ArrayList<>();
 
-  public ARGMergeJoinPredicatedAnalysis(MergeOperator pWrappedMerge) {
+  public ARGMergeJoinPredicatedAnalysis(MergeOperator pWrappedMerge, final boolean pDeleteSubgraph) {
     wrappedMerge = pWrappedMerge;
+    deleteSubgraphAfterMerge = pDeleteSubgraph;
   }
 
   // may cause problems during refinement, relink of elements may lead to non-matching abstraction formulae
@@ -74,9 +76,13 @@ public class ARGMergeJoinPredicatedAnalysis implements MergeOperator {
     if (retElement.equals(wrappedState2)) { return pState2; }
 
     ARGState mergedElement = new ARGState(retElement, null);
-    // now replace argElement2 by mergedElement in ARG
-    // deleteChildren(argElement2);
-    deleteChildren2(argElement2);
+
+    if (deleteSubgraphAfterMerge) {
+      // now replace argElement2 by mergedElement in ARG
+      // deleteChildren(argElement2);
+      deleteChildren2(argElement2);
+    }
+
     argElement2.replaceInARGWith(mergedElement);
